ad9361 parameters: Rename ad9361_parameters.h -> parameters.h
This is necessary for consistency reasons with other projects.
This header will be used by iio application, so it has to have the same
name in every project that uses iio.
